moved self heling to new file
also checking for pause and active before triggering convergence
