+++ otter/convergence/service.py
@@ -466,74 +466,6 @@ def trigger_convergence(tenant_id, group_id):
 
 
 
-class ConvergeAllGroups(MultiService, object):
-    """
-    A service that triggers convergence on all the groups every 5 mins. Only
-    one node is allowed to do this.
-    """
-
-    def __init__(self, dispatcher, kz_client, interval, log, clock,
-                 config_func):
-        super(ConvergeAllGroups, self).__init__()
-        self.disp = dispatcher
-        self.log = log
-        self.lock = self.kz_client.Lock("/selfheallock")
-        self.lock_acquired = False
-        self.clock = clock
-        self.time_range = interval - 5
-        timer = TimerService(interval, self._converge_all, config_func)
-        timer.clock = clock
-        timer.setServiceParent(self)
-
-    def stopService(self):
-        # We explicitly do not wait for convergence triggering to complete
-        # before shutting down the service. We however want to release the lock
-        # before shutting down
-        super(ConvergeAllGroups, self).stopService()
-        if self.lock_acquired:
-            return self.lock.release()
-
-    def health_check(self):
-        """
-        Return if this object has lock
-        """
-        return True, {"has_lock": self.lock_acquired}
-
-    def _perform(self):
-        d = perform(self.disp, get_groups_to_converge(config_func))
-        d.addCallback(self._setup_converges)
-        return d.addErrback(self.log.err, "self-heal-err")
-
-    def _setup_converges(self, groups):
-        wait_time = self.time_range / len(groups)
-        for i, group in enumerate(groups):
-            self.clock.callLater(
-                i * wait_time, perform, self.disp,
-                trigger_convergence(group["tenantId"], group["groupId"]))
-
-    @inlineCallbacks
-    def _converge_all(self, config_func):
-        if self.lock_acquired:
-            yield self._perform()
-        else:
-            if (yield self.lock.acquire(False, None)):
-                self.lock_acquired = True
-                yield self._perform()
-
-
-def get_groups_to_converge(time_range, config_func):
-    """
-    Get all tenant's all groups that needs convergence triggering
-    """
-    eff = Effect(GetAllValidGroups())
-
-    def should_converge(g):
-        """ Convergence enabled on tenant and ACTIVE group is not paused """
-        return (tenant_is_enabled(g["tenantId"], config_func)
-                and (not g["paused"]) and g.get("status") == "ACTIVE")
-
-    return eff.on(list(filter(should_converge)))
-
 
 class ConcurrentError(Exception):
     """Tried to run an effect concurrently when it shouldn't be."""
